Extract TestSceneMouseStates
