+++ osu.Framework.Tests/Visual/Input/TestSuiteMouseStates.cs
@@ -2,117 +2,35 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using NUnit.Framework;
-using osu.Framework.Extensions.Color4Extensions;
-using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.Sprites;
-using osu.Framework.Input.Events;
 using osu.Framework.MathUtils;
 using osu.Framework.Testing;
 using osuTK;
-using osuTK.Graphics;
 using osuTK.Input;
 
 namespace osu.Framework.Tests.Visual.Input
 {
-    public class TestSuiteMouseStates : ManualInputManagerTestSuite
+    public class TestSuiteMouseStates : ManualInputManagerTestSuite<TestSceneMouseStates>
     {
-        private readonly Box marginBox, outerMarginBox;
-        private readonly Container actionContainer;
-
-        private readonly StateTracker s1, s2;
-
-        public TestSuiteMouseStates()
-        {
-            Child = new Container
-            {
-                FillMode = FillMode.Fit,
-                FillAspectRatio = 1,
-                RelativeSizeAxes = Axes.Both,
-                Size = new Vector2(0.75f),
-                Anchor = Anchor.Centre,
-                Origin = Anchor.Centre,
-                Children = new Drawable[]
-                {
-                    new Box
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                        Colour = new Color4(1, 1, 1, 0.2f),
-                    },
-                    s1 = new StateTracker(1),
-                    new Container
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                        Children = new Drawable[]
-                        {
-                            outerMarginBox = new Box
-                            {
-                                RelativeSizeAxes = Axes.Both,
-                                Anchor = Anchor.Centre,
-                                Origin = Anchor.Centre,
-                                Size = new Vector2(0.9f),
-                                Colour = Color4.SkyBlue.Opacity(0.1f),
-                            },
-                            actionContainer = new Container
-                            {
-                                RelativeSizeAxes = Axes.Both,
-                                Size = new Vector2(0.6f),
-                                Anchor = Anchor.Centre,
-                                Origin = Anchor.Centre,
-                                Children = new Drawable[]
-                                {
-                                    new Box
-                                    {
-                                        RelativeSizeAxes = Axes.Both,
-                                        Colour = new Color4(1, 1, 1, 0.2f),
-                                    },
-                                    marginBox = new Box
-                                    {
-                                        RelativeSizeAxes = Axes.Both,
-                                        Anchor = Anchor.Centre,
-                                        Origin = Anchor.Centre,
-                                        Size = new Vector2(0.8f),
-                                        Colour = Color4.SkyBlue.Opacity(0.1f),
-                                    },
-                                    s2 = new DraggableStateTracker(2),
-                                }
-                            }
-                        }
-                    }
-                }
-            };
-        }
-
         protected override void LoadComplete()
         {
-            ((Container)InputManager.Parent).Add(new StateTracker(0));
+            ((Container)InputManager.Parent).Add(new TestSceneMouseStates.StateTracker(0));
         }
 
         private void initTestScene()
         {
-            eventCounts1.Clear();
-            eventCounts2.Clear();
+            TestScene.EventCounts1.Clear();
+            TestScene.EventCounts2.Clear();
             // InitialMousePosition cannot be used here because the event counters should be resetted after the initial mouse move.
-            AddStep("move mouse to center", () => InputManager.MoveMouseTo(actionContainer));
+            AddStep("move mouse to center", () => InputManager.MoveMouseTo(TestScene.ActionContainer));
             AddStep("reset event counters", () =>
             {
-                s1.Reset();
-                s2.Reset();
+                TestScene.S1.Reset();
+                TestScene.S2.Reset();
             });
         }
-
-        private static readonly Type move = typeof(MouseMoveEvent);
-        private static readonly Type scroll = typeof(ScrollEvent);
-        private static readonly Type mouse_down = typeof(MouseDownEvent);
-        private static readonly Type mouse_up = typeof(MouseUpEvent);
-        private static readonly Type drag_start = typeof(DragStartEvent);
-        private static readonly Type drag_end = typeof(DragEndEvent);
-        private static readonly Type click = typeof(ClickEvent);
-        private static readonly Type double_click = typeof(DoubleClickEvent);
 
         [Test]
         public void BasicScroll()
@@ -120,12 +38,12 @@ namespace osu.Framework.Tests.Visual.Input
             initTestScene();
 
             AddStep("scroll some", () => InputManager.ScrollBy(new Vector2(-1, 1)));
-            checkEventCount(move);
-            checkEventCount(scroll, 1);
+            checkEventCount(TestSceneMouseStates.MOVE);
+            checkEventCount(TestSceneMouseStates.SCROLL, 1);
             checkLastScrollDelta(new Vector2(-1, 1));
 
             AddStep("scroll some", () => InputManager.ScrollBy(new Vector2(1, -1)));
-            checkEventCount(scroll, 1);
+            checkEventCount(TestSceneMouseStates.SCROLL, 1);
             checkLastScrollDelta(new Vector2(1, -1));
         }
 
@@ -134,26 +52,26 @@ namespace osu.Framework.Tests.Visual.Input
         {
             initTestScene();
 
-            AddStep("push move", () => InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.TopLeft));
-            checkEventCount(move, 1);
-            checkEventCount(scroll);
+            AddStep("push move", () => InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.TopLeft));
+            checkEventCount(TestSceneMouseStates.MOVE, 1);
+            checkEventCount(TestSceneMouseStates.SCROLL);
 
-            AddStep("push move", () => InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.TopRight));
-            checkEventCount(move, 1);
-            checkEventCount(scroll);
-            checkLastPositionDelta(() => marginBox.ScreenSpaceDrawQuad.Width);
+            AddStep("push move", () => InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.TopRight));
+            checkEventCount(TestSceneMouseStates.MOVE, 1);
+            checkEventCount(TestSceneMouseStates.SCROLL);
+            checkLastPositionDelta(() => TestScene.MarginBox.ScreenSpaceDrawQuad.Width);
 
-            AddStep("push move", () => InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.BottomRight));
-            checkEventCount(move, 1);
-            checkLastPositionDelta(() => marginBox.ScreenSpaceDrawQuad.Height);
+            AddStep("push move", () => InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.BottomRight));
+            checkEventCount(TestSceneMouseStates.MOVE, 1);
+            checkLastPositionDelta(() => TestScene.MarginBox.ScreenSpaceDrawQuad.Height);
 
             AddStep("push two moves", () =>
             {
-                InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.TopLeft);
-                InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.BottomLeft);
+                InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.TopLeft);
+                InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.BottomLeft);
             });
-            checkEventCount(move, 2);
-            checkLastPositionDelta(() => Vector2.Distance(marginBox.ScreenSpaceDrawQuad.TopLeft, marginBox.ScreenSpaceDrawQuad.BottomLeft));
+            checkEventCount(TestSceneMouseStates.MOVE, 2);
+            checkLastPositionDelta(() => Vector2.Distance(TestScene.MarginBox.ScreenSpaceDrawQuad.TopLeft, TestScene.MarginBox.ScreenSpaceDrawQuad.BottomLeft));
         }
 
         [Test]
@@ -162,16 +80,16 @@ namespace osu.Framework.Tests.Visual.Input
             initTestScene();
 
             AddStep("press left button", () => InputManager.PressButton(MouseButton.Left));
-            checkEventCount(mouse_down, 1);
+            checkEventCount(TestSceneMouseStates.MOUSE_DOWN, 1);
 
             AddStep("press right button", () => InputManager.PressButton(MouseButton.Right));
-            checkEventCount(mouse_down, 1);
+            checkEventCount(TestSceneMouseStates.MOUSE_DOWN, 1);
 
             AddStep("release left button", () => InputManager.ReleaseButton(MouseButton.Left));
-            checkEventCount(mouse_up, 1);
+            checkEventCount(TestSceneMouseStates.MOUSE_UP, 1);
 
             AddStep("release right button", () => InputManager.ReleaseButton(MouseButton.Right));
-            checkEventCount(mouse_up, 1);
+            checkEventCount(TestSceneMouseStates.MOUSE_UP, 1);
 
             AddStep("press three buttons", () =>
             {
@@ -179,7 +97,7 @@ namespace osu.Framework.Tests.Visual.Input
                 InputManager.PressButton(MouseButton.Right);
                 InputManager.PressButton(MouseButton.Button1);
             });
-            checkEventCount(mouse_down, 3);
+            checkEventCount(TestSceneMouseStates.MOUSE_DOWN, 3);
 
             AddStep("Release mouse buttons", () =>
             {
@@ -187,7 +105,7 @@ namespace osu.Framework.Tests.Visual.Input
                 InputManager.ReleaseButton(MouseButton.Right);
                 InputManager.ReleaseButton(MouseButton.Button1);
             });
-            checkEventCount(mouse_up, 3);
+            checkEventCount(TestSceneMouseStates.MOUSE_UP, 3);
 
             AddStep("press two buttons", () =>
             {
@@ -196,13 +114,13 @@ namespace osu.Framework.Tests.Visual.Input
                 InputManager.PressButton(MouseButton.Right);
             });
 
-            checkEventCount(mouse_down, 2);
-            checkEventCount(mouse_up, 1);
+            checkEventCount(TestSceneMouseStates.MOUSE_DOWN, 2);
+            checkEventCount(TestSceneMouseStates.MOUSE_UP, 1);
 
             AddStep("release", () => InputManager.ReleaseButton(MouseButton.Right));
 
-            checkEventCount(move);
-            checkEventCount(mouse_up, 1);
+            checkEventCount(TestSceneMouseStates.MOVE);
+            checkEventCount(TestSceneMouseStates.MOUSE_UP, 1);
         }
 
         [Test]
@@ -211,15 +129,15 @@ namespace osu.Framework.Tests.Visual.Input
             initTestScene();
 
             AddStep("press left button", () => InputManager.PressButton(MouseButton.Left));
-            checkEventCount(mouse_down, 1);
+            checkEventCount(TestSceneMouseStates.MOUSE_DOWN, 1);
             checkIsDragged(false);
 
-            AddStep("move bottom left", () => InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.BottomLeft));
-            checkEventCount(drag_start, 1);
+            AddStep("move bottom left", () => InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.BottomLeft));
+            checkEventCount(TestSceneMouseStates.DRAG_START, 1);
             checkIsDragged(true);
 
             AddStep("release left button", () => InputManager.ReleaseButton(MouseButton.Left));
-            checkEventCount(mouse_up, 1);
+            checkEventCount(TestSceneMouseStates.MOUSE_UP, 1);
             checkIsDragged(false);
         }
 
@@ -228,24 +146,24 @@ namespace osu.Framework.Tests.Visual.Input
         {
             initTestScene();
 
-            AddStep("push move", () => InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.BottomLeft));
-            checkEventCount(move, 1);
+            AddStep("push move", () => InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.BottomLeft));
+            checkEventCount(TestSceneMouseStates.MOVE, 1);
 
             AddStep("push move and scroll", () =>
             {
-                InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.Centre);
+                InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.Centre);
                 InputManager.ScrollBy(new Vector2(1, 2));
             });
 
-            checkEventCount(move, 1);
-            checkEventCount(scroll, 1);
+            checkEventCount(TestSceneMouseStates.MOVE, 1);
+            checkEventCount(TestSceneMouseStates.SCROLL, 1);
             checkLastScrollDelta(new Vector2(1, 2));
-            checkLastPositionDelta(() => Vector2.Distance(marginBox.ScreenSpaceDrawQuad.BottomLeft, marginBox.ScreenSpaceDrawQuad.Centre));
+            checkLastPositionDelta(() => Vector2.Distance(TestScene.MarginBox.ScreenSpaceDrawQuad.BottomLeft, TestScene.MarginBox.ScreenSpaceDrawQuad.Centre));
 
             AddStep("Move mouse to out of bounds", () => InputManager.MoveMouseTo(Vector2.Zero));
 
-            checkEventCount(move);
-            checkEventCount(scroll);
+            checkEventCount(TestSceneMouseStates.MOVE);
+            checkEventCount(TestSceneMouseStates.SCROLL);
 
             AddStep("Move mouse", () =>
             {
@@ -254,8 +172,8 @@ namespace osu.Framework.Tests.Visual.Input
             });
 
             // outside the bounds so should not increment.
-            checkEventCount(move);
-            checkEventCount(scroll);
+            checkEventCount(TestSceneMouseStates.MOVE);
+            checkEventCount(TestSceneMouseStates.SCROLL);
         }
 
         [Test]
@@ -264,36 +182,36 @@ namespace osu.Framework.Tests.Visual.Input
             initTestScene();
 
             // mouseDown on a non-draggable -> mouseUp on a distant position: drag-clicking
-            AddStep("move mouse", () => InputManager.MoveMouseTo(outerMarginBox.ScreenSpaceDrawQuad.TopLeft));
+            AddStep("move mouse", () => InputManager.MoveMouseTo(TestScene.OuterMarginBox.ScreenSpaceDrawQuad.TopLeft));
             AddStep("press left button", () => InputManager.PressButton(MouseButton.Left));
-            checkEventCount(drag_start);
-            AddStep("drag non-draggable", () => InputManager.MoveMouseTo(marginBox));
-            checkEventCount(drag_start, 1, true);
+            checkEventCount(TestSceneMouseStates.DRAG_START);
+            AddStep("drag non-draggable", () => InputManager.MoveMouseTo(TestScene.MarginBox));
+            checkEventCount(TestSceneMouseStates.DRAG_START, 1, true);
             AddStep("release left button", () => InputManager.ReleaseButton(MouseButton.Left));
-            checkEventCount(click, 1, true);
-            checkEventCount(drag_end);
+            checkEventCount(TestSceneMouseStates.CLICK, 1, true);
+            checkEventCount(TestSceneMouseStates.DRAG_END);
 
             // mouseDown on a draggable -> mouseUp on the original position: no drag-clicking
-            AddStep("move mouse", () => InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.TopLeft));
+            AddStep("move mouse", () => InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.TopLeft));
             AddStep("press left button", () => InputManager.PressButton(MouseButton.Left));
-            AddStep("drag draggable", () => InputManager.MoveMouseTo(outerMarginBox.ScreenSpaceDrawQuad.BottomRight));
-            checkEventCount(drag_start, 1);
+            AddStep("drag draggable", () => InputManager.MoveMouseTo(TestScene.OuterMarginBox.ScreenSpaceDrawQuad.BottomRight));
+            checkEventCount(TestSceneMouseStates.DRAG_START, 1);
             checkIsDragged(true);
-            AddStep("return mouse position", () => InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.TopLeft));
+            AddStep("return mouse position", () => InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.TopLeft));
             checkIsDragged(true);
-            checkEventCount(drag_end);
+            checkEventCount(TestSceneMouseStates.DRAG_END);
             AddStep("release left button", () => InputManager.ReleaseButton(MouseButton.Left));
-            checkEventCount(click);
-            checkEventCount(drag_end, 1);
+            checkEventCount(TestSceneMouseStates.CLICK);
+            checkEventCount(TestSceneMouseStates.DRAG_END, 1);
             checkIsDragged(false);
 
             // mouseDown on a draggable -> mouseUp on a distant position: no drag-clicking
             AddStep("press left button", () => InputManager.PressButton(MouseButton.Left));
-            AddStep("drag draggable", () => InputManager.MoveMouseTo(marginBox.ScreenSpaceDrawQuad.BottomRight));
+            AddStep("drag draggable", () => InputManager.MoveMouseTo(TestScene.MarginBox.ScreenSpaceDrawQuad.BottomRight));
             AddStep("release left button", () => InputManager.ReleaseButton(MouseButton.Left));
-            checkEventCount(drag_start, 1);
-            checkEventCount(drag_end, 1);
-            checkEventCount(click);
+            checkEventCount(TestSceneMouseStates.DRAG_START, 1);
+            checkEventCount(TestSceneMouseStates.DRAG_END, 1);
+            checkEventCount(TestSceneMouseStates.CLICK);
         }
 
         [Test]
@@ -303,18 +221,18 @@ namespace osu.Framework.Tests.Visual.Input
 
             waitDoubleClickTime();
             AddStep("click", () => InputManager.Click(MouseButton.Left));
-            checkEventCount(click, 1);
+            checkEventCount(TestSceneMouseStates.CLICK, 1);
             waitDoubleClickTime();
             AddStep("click", () => InputManager.Click(MouseButton.Left));
-            checkEventCount(click, 1);
+            checkEventCount(TestSceneMouseStates.CLICK, 1);
             waitDoubleClickTime();
             AddStep("double click", () =>
             {
                 InputManager.Click(MouseButton.Left);
                 InputManager.Click(MouseButton.Left);
             });
-            checkEventCount(click, 1);
-            checkEventCount(double_click, 1);
+            checkEventCount(TestSceneMouseStates.CLICK, 1);
+            checkEventCount(TestSceneMouseStates.DOUBLE_CLICK, 1);
             waitDoubleClickTime();
             AddStep("triple click", () =>
             {
@@ -322,8 +240,8 @@ namespace osu.Framework.Tests.Visual.Input
                 InputManager.Click(MouseButton.Left);
                 InputManager.Click(MouseButton.Left);
             });
-            checkEventCount(click, 2);
-            checkEventCount(double_click, 1);
+            checkEventCount(TestSceneMouseStates.CLICK, 2);
+            checkEventCount(TestSceneMouseStates.DOUBLE_CLICK, 1);
 
             waitDoubleClickTime();
             AddStep("click then mouse down", () =>
@@ -331,22 +249,22 @@ namespace osu.Framework.Tests.Visual.Input
                 InputManager.Click(MouseButton.Left);
                 InputManager.PressButton(MouseButton.Left);
             });
-            checkEventCount(click, 1);
-            checkEventCount(double_click, 1);
+            checkEventCount(TestSceneMouseStates.CLICK, 1);
+            checkEventCount(TestSceneMouseStates.DOUBLE_CLICK, 1);
             AddStep("mouse up", () => InputManager.ReleaseButton(MouseButton.Left));
-            checkEventCount(click);
-            checkEventCount(double_click);
+            checkEventCount(TestSceneMouseStates.CLICK);
+            checkEventCount(TestSceneMouseStates.DOUBLE_CLICK);
 
             waitDoubleClickTime();
             AddStep("double click drag", () =>
             {
                 InputManager.Click(MouseButton.Left);
                 InputManager.PressButton(MouseButton.Left);
-                InputManager.MoveMouseTo(outerMarginBox.ScreenSpaceDrawQuad.TopLeft);
+                InputManager.MoveMouseTo(TestScene.OuterMarginBox.ScreenSpaceDrawQuad.TopLeft);
             });
-            checkEventCount(click, 1);
-            checkEventCount(double_click, 1);
-            checkEventCount(drag_start, 1);
+            checkEventCount(TestSceneMouseStates.CLICK, 1);
+            checkEventCount(TestSceneMouseStates.DOUBLE_CLICK, 1);
+            checkEventCount(TestSceneMouseStates.DRAG_START, 1);
         }
 
         [Test]
@@ -355,13 +273,13 @@ namespace osu.Framework.Tests.Visual.Input
             initTestScene();
 
             AddStep("right down", () => InputManager.PressButton(MouseButton.Right));
-            checkEventCount(mouse_down, 1);
-            AddStep("move away", () => InputManager.MoveMouseTo(outerMarginBox.ScreenSpaceDrawQuad.TopLeft));
+            checkEventCount(TestSceneMouseStates.MOUSE_DOWN, 1);
+            AddStep("move away", () => InputManager.MoveMouseTo(TestScene.OuterMarginBox.ScreenSpaceDrawQuad.TopLeft));
             AddStep("left click", () => InputManager.Click(MouseButton.Left));
-            checkEventCount(mouse_down, 1, true);
-            checkEventCount(mouse_up, 1, true);
+            checkEventCount(TestSceneMouseStates.MOUSE_DOWN, 1, true);
+            checkEventCount(TestSceneMouseStates.MOUSE_UP, 1, true);
             AddStep("right up", () => InputManager.ReleaseButton(MouseButton.Right));
-            checkEventCount(mouse_up, 1);
+            checkEventCount(TestSceneMouseStates.MOUSE_UP, 1);
         }
 
         private void waitDoubleClickTime()
@@ -369,13 +287,10 @@ namespace osu.Framework.Tests.Visual.Input
             AddWaitStep("wait to don't double click", 2);
         }
 
-        private readonly Dictionary<Type, int> eventCounts1 = new Dictionary<Type, int>(),
-                                               eventCounts2 = new Dictionary<Type, int>();
-
         private void checkEventCount(Type type, int change = 0, bool outer = false)
         {
-            eventCounts1.TryGetValue(type, out var count1);
-            eventCounts2.TryGetValue(type, out var count2);
+            TestScene.EventCounts1.TryGetValue(type, out var count1);
+            TestScene.EventCounts2.TryGetValue(type, out var count2);
 
             if (outer)
             {
@@ -384,244 +299,25 @@ namespace osu.Framework.Tests.Visual.Input
             else
             {
                 // those types are handled by state tracker 2
-                if (!new[] { drag_start, drag_end, click, double_click }.Contains(type))
+                if (!new[] { TestSceneMouseStates.DRAG_START, TestSceneMouseStates.DRAG_END, TestSceneMouseStates.CLICK, TestSceneMouseStates.DOUBLE_CLICK }.Contains(type))
                     count1 += change;
                 count2 += change;
             }
 
-            AddAssert($"{type.Name} count {count1}, {count2}", () => s1.CounterFor(type).Count == count1 && s2.CounterFor(type).Count == count2);
+            AddAssert($"{type.Name} count {count1}, {count2}", () => TestScene.S1.CounterFor(type).Count == count1 && TestScene.S2.CounterFor(type).Count == count2);
 
-            eventCounts1[type] = count1;
-            eventCounts2[type] = count2;
+            TestScene.EventCounts1[type] = count1;
+            TestScene.EventCounts2[type] = count2;
         }
 
         private void checkLastPositionDelta(Func<float> expected) => AddAssert("correct position delta", () =>
-            Precision.AlmostEquals(s1.LastDelta.Length, expected()) &&
-            Precision.AlmostEquals(s2.LastDelta.Length, expected()));
+            Precision.AlmostEquals(TestScene.S1.LastDelta.Length, expected()) &&
+            Precision.AlmostEquals(TestScene.S2.LastDelta.Length, expected()));
 
         private void checkLastScrollDelta(Vector2 expected) => AddAssert("correct scroll delta", () =>
-            Precision.AlmostEquals(s1.LastScrollDelta, expected) &&
-            Precision.AlmostEquals(s2.LastScrollDelta, expected));
+            Precision.AlmostEquals(TestScene.S1.LastScrollDelta, expected) &&
+            Precision.AlmostEquals(TestScene.S2.LastScrollDelta, expected));
 
-        private void checkIsDragged(bool isDragged) => AddAssert(isDragged ? "dragged" : "not dragged", () => s2.IsDragged == isDragged);
-
-        public class StateTracker : Container
-        {
-            private readonly SpriteText keyboard;
-            private readonly SpriteText mouse;
-            private readonly SpriteText source;
-            protected readonly FillFlowContainer TextContainer;
-
-            public EventCounter CounterFor(Type type) => counterLookup[type];
-            public Vector2 LastDelta { get; private set; }
-            public Vector2 LastScrollDelta { get; private set; }
-
-            private readonly Dictionary<Type, EventCounter> counterLookup = new Dictionary<Type, EventCounter>();
-
-            public StateTracker(int number)
-            {
-                RelativeSizeAxes = Axes.Both;
-                Margin = new MarginPadding(5);
-                Children = new Drawable[]
-                {
-                    TextContainer = new FillFlowContainer
-                    {
-                        Direction = FillDirection.Vertical,
-                        Children = new Drawable[]
-                        {
-                            source = new SmallText(),
-                            keyboard = new SmallText(),
-                            mouse = new SmallText(),
-                            addCounter(typeof(ScrollEvent)),
-                            addCounter(typeof(MouseMoveEvent)),
-                            addCounter(typeof(DragStartEvent)),
-                            addCounter(typeof(DragEndEvent)),
-                            addCounter(typeof(MouseDownEvent)),
-                            addCounter(typeof(MouseUpEvent)),
-                            addCounter(typeof(ClickEvent)),
-                            addCounter(typeof(DoubleClickEvent))
-                        }
-                    },
-                    new BoundedCursorContainer(number)
-                };
-            }
-
-            protected override bool Handle(UIEvent e)
-            {
-                var type = e.GetType();
-                if (!counterLookup.TryGetValue(type, out var counter))
-                    return base.Handle(e);
-
-                ++counter.Count;
-
-                switch (e)
-                {
-                    case MouseMoveEvent mouseMove:
-                        LastDelta = mouseMove.ScreenSpaceMousePosition - mouseMove.ScreenSpaceLastMousePosition;
-                        break;
-
-                    case ScrollEvent scroll:
-                        LastScrollDelta = scroll.ScrollDelta;
-                        break;
-                }
-
-                return type == click || type == double_click;
-            }
-
-            private EventCounter addCounter(Type type)
-            {
-                var counter = new EventCounter(type);
-                counterLookup.Add(type, counter);
-                return counter;
-            }
-
-            public void Reset()
-            {
-                foreach (var kvp in counterLookup)
-                    kvp.Value.Reset();
-
-                LastDelta = Vector2.Zero;
-                LastScrollDelta = Vector2.Zero;
-            }
-
-            protected override void Update()
-            {
-                base.Update();
-
-                var inputManager = GetContainingInputManager();
-
-                if (inputManager != null)
-                {
-                    var state = inputManager.CurrentState;
-
-                    source.Text = inputManager.ToString();
-                    keyboard.Text = state.Keyboard.ToString();
-                    mouse.Text = state.Mouse.ToString();
-                }
-            }
-
-            public class SmallText : SpriteText
-            {
-                public SmallText()
-                {
-                    Font = new FontUsage(size: 14);
-                }
-            }
-
-            public class EventCounter : CompositeDrawable
-            {
-                private int count;
-                private readonly SpriteText text;
-
-                public EventCounter(Type eventType)
-                {
-                    AutoSizeAxes = Axes.Both;
-                    InternalChild = text = new SmallText();
-
-                    Name = eventType.Name.Replace("Event", "");
-
-                    Reset();
-                }
-
-                public int Count
-                {
-                    get => count;
-                    set
-                    {
-                        count = value;
-                        text.Text = $"{Name}: {Count}";
-                    }
-                }
-
-                public void Reset()
-                {
-                    Count = 0;
-                }
-            }
-
-            public class BoundedCursorContainer : Container
-            {
-                private readonly Circle circle;
-
-                public BoundedCursorContainer(int number)
-                {
-                    RelativeSizeAxes = Axes.Both;
-
-                    Child = new Container
-                    {
-                        Size = new Vector2(5),
-                        Origin = Anchor.Centre,
-                        Children = new Drawable[]
-                        {
-                            new Circle
-                            {
-                                Origin = Anchor.Centre,
-                                Size = new Vector2(40),
-                                Alpha = 0.1f,
-                            },
-                            new Container
-                            {
-                                RelativeSizeAxes = Axes.Both,
-                                Origin = Anchor.Centre,
-                                X = -5 + number * 5,
-                                Y = -8,
-                                Child = circle = new Circle
-                                {
-                                    RelativeSizeAxes = Axes.Both,
-                                }
-                            }
-                        }
-                    };
-                }
-
-                protected override bool OnMouseMove(MouseMoveEvent e)
-                {
-                    Child.MoveTo(e.MousePosition, 100, Easing.OutQuint);
-                    return base.OnMouseMove(e);
-                }
-
-                protected override bool OnScroll(ScrollEvent e)
-                {
-                    circle.MoveTo(circle.Position - e.ScrollDelta * 10).MoveTo(Vector2.Zero, 500, Easing.OutQuint);
-                    return base.OnScroll(e);
-                }
-
-                protected override bool OnMouseDown(MouseDownEvent e)
-                {
-                    adjustForMouseDown(e);
-                    return base.OnMouseDown(e);
-                }
-
-                protected override bool OnMouseUp(MouseUpEvent e)
-                {
-                    adjustForMouseDown(e);
-                    return base.OnMouseUp(e);
-                }
-
-                private void adjustForMouseDown(MouseEvent e)
-                {
-                    circle.FadeColour(e.HasAnyButtonPressed ? Color4.Green.Lighten((e.PressedButtons.Count() - 1) * 0.3f) : Color4.White, 50);
-                }
-            }
-        }
-
-        public class DraggableStateTracker : StateTracker
-        {
-            private readonly SmallText dragStatus;
-
-            public DraggableStateTracker(int number)
-                : base(number)
-            {
-                TextContainer.Add(dragStatus = new SmallText());
-            }
-
-            protected override bool Handle(UIEvent e) => base.Handle(e) || e is DragStartEvent;
-
-            protected override void Update()
-            {
-                base.Update();
-                dragStatus.Text = $"IsDragged = {IsDragged}";
-            }
-        }
+        private void checkIsDragged(bool isDragged) => AddAssert(isDragged ? "dragged" : "not dragged", () => TestScene.S2.IsDragged == isDragged);
     }
 }
